move useProgram back to Painter
