+++ js/render/painter.js
@@ -12,6 +12,8 @@ const VertexArrayObject = require('./vertex_array_object');
 const RasterBoundsArray = require('../data/raster_bounds_array');
 const PosArray = require('../data/pos_array');
 const ProgramConfiguration = require('../data/program_configuration');
+const shaders = require('mapbox-gl-shaders');
+const assert = require('assert');
 
 const draw = {
     symbol: require('./draw_symbol'),
@@ -50,7 +52,6 @@ class Painter {
         this.lineWidthRange = gl.getParameter(gl.ALIASED_LINE_WIDTH_RANGE);
 
         this.basicFillProgramConfiguration = ProgramConfiguration.createStatic(['color', 'opacity']);
-        this.emptyProgramConfiguration = ProgramConfiguration.createStatic([]);
     }
 
     /*
@@ -356,18 +357,57 @@ class Painter {
         }
     }
 
+    createProgram(name, showOverdraw, gl, configuration) {
+        const program = gl.createProgram();
+        const definition = shaders[name];
+
+        let definesSource = '#define MAPBOX_GL_JS;\n';
+        if (showOverdraw) {
+            definesSource += '#define OVERDRAW_INSPECTOR;\n';
+        }
+
+        const fragmentShader = gl.createShader(gl.FRAGMENT_SHADER);
+        gl.shaderSource(fragmentShader, applyPragmas(definesSource + definition.fragmentSource, configuration.fragmentPragmas));
+        gl.compileShader(fragmentShader);
+        assert(gl.getShaderParameter(fragmentShader, gl.COMPILE_STATUS), gl.getShaderInfoLog(fragmentShader));
+        gl.attachShader(program, fragmentShader);
+
+        const vertexShader = gl.createShader(gl.VERTEX_SHADER);
+        gl.shaderSource(vertexShader, applyPragmas(definesSource + shaders.util + definition.vertexSource, configuration.vertexPragmas));
+        gl.compileShader(vertexShader);
+        assert(gl.getShaderParameter(vertexShader, gl.COMPILE_STATUS), gl.getShaderInfoLog(vertexShader));
+        gl.attachShader(program, vertexShader);
+
+        gl.linkProgram(program);
+        assert(gl.getProgramParameter(program, gl.LINK_STATUS), gl.getProgramInfoLog(program));
+
+        const numAttributes = gl.getProgramParameter(program, gl.ACTIVE_ATTRIBUTES);
+        const result = {program, numAttributes};
+
+        for (let i = 0; i < numAttributes; i++) {
+            const attribute = gl.getActiveAttrib(program, i);
+            result[attribute.name] = gl.getAttribLocation(program, attribute.name);
+        }
+        const numUniforms = gl.getProgramParameter(program, gl.ACTIVE_UNIFORMS);
+        for (let i = 0; i < numUniforms; i++) {
+            const uniform = gl.getActiveUniform(program, i);
+            result[uniform.name] = gl.getUniformLocation(program, uniform.name);
+        }
+        return result;
+    }
+
     _createProgramCached(name, programConfiguration) {
         this.cache = this.cache || {};
         const key = `${name}${programConfiguration.cacheKey}${!!this._showOverdrawInspector}`;
         if (!this.cache[key]) {
-            this.cache[key] = programConfiguration.createProgram(name, this._showOverdrawInspector, this.gl);
+            this.cache[key] = this.createProgram(name, this._showOverdrawInspector, this.gl, programConfiguration);
         }
         return this.cache[key];
     }
 
     useProgram(name, programConfiguration) {
         const gl = this.gl;
-        const nextProgram = this._createProgramCached(name, programConfiguration || this.emptyProgramConfiguration);
+        const nextProgram = this._createProgramCached(name, programConfiguration || {});
 
         if (this.currentProgram !== nextProgram) {
             gl.useProgram(nextProgram.program);
@@ -376,6 +416,15 @@ class Painter {
 
         return nextProgram;
     }
+}
+
+function applyPragmas(source, pragmas) {
+    return source.replace(/#pragma mapbox: ([\w]+) ([\w]+) ([\w]+) ([\w]+)/g, (match, operation, precision, type, name) => {
+        return pragmas[name][operation]
+            .join('\n')
+            .replace(/{type}/g, type)
+            .replace(/{precision}/g, precision);
+    });
 }
 
 module.exports = Painter;
